+++ bin/alethia-reth/src/main.rs
@@ -1,5 +1,5 @@
 //! Rust Taiko node (alethia-reth) binary executable.
-use alethia_reth_cli::{TaikoCli, chainspec::TaikoChainSpecParser};
+use alethia_reth_cli::{TaikoCli, TaikoCliExtArgs, chainspec::TaikoChainSpecParser};
 use alethia_reth_node::{
     TaikoNode,
     rpc::eth::{
@@ -7,7 +7,7 @@ use alethia_reth_node::{
         eth::{TaikoExt, TaikoExtApiServer},
     },
 };
-use reth::{args::RessArgs, builder::NodeHandle, ress::install_ress_subprotocol};
+use reth::{builder::NodeHandle, ress::install_ress_subprotocol};
 use reth_rpc::eth::{EthApiTypes, RpcNodeCore};
 use tracing::info;
 
@@ -20,8 +20,9 @@ fn main() {
         unsafe { std::env::set_var("RUST_BACKTRACE", "1") };
     }
 
-    if let Err(err) = TaikoCli::<TaikoChainSpecParser, RessArgs>::parse_args().run(
-        async move |builder, ress_args| {
+    if let Err(err) = TaikoCli::<TaikoChainSpecParser, TaikoCliExtArgs>::parse_args().run(
+        async move |builder, ext_args| {
+            let TaikoCliExtArgs { ress, .. } = ext_args;
             info!(target: "reth::taiko::cli", "Launching Taiko node");
             let NodeHandle { node, node_exit_future } = builder
                 .node(TaikoNode)
@@ -47,9 +48,9 @@ fn main() {
                 .await?;
 
             // Install ress subprotocol.
-            if ress_args.enabled {
+            if ress.enabled {
                 install_ress_subprotocol(
-                    ress_args,
+                    ress,
                     node.provider,
                     node.evm_config,
                     node.network,